[#21] Refactor: account 유효성 검사 util 함수로 분리
