+++ src/utils/index.js
@@ -1,6 +1,2 @@
-import { USER_TOKEN_KEY } from 'constants';
-
-export const getUserToken = () => {
-  const userToken = localStorage.getItem(USER_TOKEN_KEY);
-  return userToken ? `Bearer ${userToken}` : null;
-};
+export * from './validate';
+export * from './storage';
